refactor and add enable radiance option to energyplussetup
