+++ Makefile
@@ -1,0 +1,5 @@
+test:
+	pytest test
+
+clean:
+	rm -f *.csv *.end *.mtr *.mdd *.bnd *.dxf *.audit *.err *.eso *.json *.rdd *.shd *.sql *.htm *.eio *.mtd
